Oliver E suggestions pt.2
